backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
dnspython==1.16.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pymongo==3.12.0
python-dotenv==1.2.1
Werkzeug==3.1.5
